add logging

+++ pkg/postgrespool/config.go
@@ -8,6 +8,6 @@ type Config struct {
 	Host         string        `env:"POSTGRES_HOST" env-default:"localhost"`
 	Port         string        `env:"POSTGRES_PORT" env-default:"5432"`
 	Database     string        `env:"POSTGRES_DATABASE" env-default:"postgres"`
-	MaxAttemps   int64         `env:"POSTGRES_ATTEMPS" env-default:"5"`
+	MaxAttemps   int           `env:"POSTGRES_ATTEMPS" env-default:"5"`
 	DelayAttemps time.Duration `env:"POSTGRES_ATTEMPS_DELAY" env-default:"3s"`
 }
